Calculate journey specifications

Uses scheduled times in the timetable to calculate how long it should take to travel along the track. Calculates journey times between stations specified in the spec_stations parameter.

Usage

calculate_journey_specifications(timetable, spec_stations)

Arguments

timetable

Data frame containing timetable data. Strict input validation is performed. Must be a data frame with the following columns:

train_headercharacter() Train identifier

dt_originlubridate::POSIXct() Datetime at which the train originated

geocharacter() Name of event location

eventcharacter() Type of event, must only contain the following values: Pass, Arrive, Depart, Originate, Terminate

wttlubridate::POSIXct() Scheduled datetime of event

tlubridate::POSIXct() Actual datetime of event

delaynumeric() Difference between actual time and scheduled time of event

allownumeric() Timetabled delay allowance

spec_stations

List of stations for which to calculate journey times. Calculates the total journey time from the first station in the list to the last station in the list, as well as every component of that journey.

Value

Journey times as specified by spec_stations. A data frame with 10 columns:

train_headerTrain identifier

dt_originDatetime at which the train originated

wttScheduled datetime of event

groupCalling pattern

xiStart of journey component

xkEnd of journey component

durationDuration of journey component

typeType of component, either moving for sections in-between stations, or dwell when the component represents a stop at a station

jComponent ID

T_journeyTotal journey time across all components

See also

Examples

data(timetable_subset)
timetable_subset
#> # A tibble: 792 × 12
#>    train_header dt_origin           geo    event  wtt                
#>    <chr>        <dttm>              <chr>  <chr>  <dttm>             
#>  1 108H         2000-01-01 12:55:00 geo5   Arrive 2000-01-01 13:34:00
#>  2 108H         2000-01-01 12:55:00 geo5   Depart 2000-01-01 13:35:30
#>  3 108H         2000-01-01 12:55:00 geo6   Pass   2000-01-01 13:37:30
#>  4 108H         2000-01-01 12:55:00 geo110 Arrive 2000-01-01 13:39:00
#>  5 108H         2000-01-01 12:55:00 geo110 Depart 2000-01-01 13:39:30
#>  6 108H         2000-01-01 12:55:00 geo111 Arrive 2000-01-01 13:42:00
#>  7 108H         2000-01-01 12:55:00 geo111 Depart 2000-01-01 13:42:30
#>  8 108H         2000-01-01 12:55:00 geo112 Arrive 2000-01-01 13:44:30
#>  9 108H         2000-01-01 12:55:00 geo112 Depart 2000-01-01 13:45:00
#> 10 108H         2000-01-01 12:55:00 geo7   Arrive 2000-01-01 13:48:00
#> # ℹ 782 more rows
#> # ℹ 7 more variables: t <dttm>, delay <dbl>, allow <dbl>, allow_perf <dbl>,
#> #   allow_path <dbl>, allow_eng <dbl>, group <chr>
# spec_stations contains stations that completely encompass the Centrix track
spec_stations <- list("geo6", "geo110", "geo111", "geo112", "geo7")
calculate_journey_specifications(timetable_subset, spec_stations)
#> # A tibble: 281 × 10
#>    train_header dt_origin           wtt                 group        xi    xk   
#>    <chr>        <dttm>              <dttm>              <chr>        <chr> <chr>
#>  1 108H         2000-01-01 12:55:00 2000-01-01 13:39:00 stopping-all geo6  geo1…
#>  2 108H         2000-01-01 12:55:00 2000-01-01 13:39:30 stopping-all geo1… geo1…
#>  3 108H         2000-01-01 12:55:00 2000-01-01 13:42:00 stopping-all geo1… geo1…
#>  4 108H         2000-01-01 12:55:00 2000-01-01 13:42:30 stopping-all geo1… geo1…
#>  5 108H         2000-01-01 12:55:00 2000-01-01 13:44:30 stopping-all geo1… geo1…
#>  6 108H         2000-01-01 12:55:00 2000-01-01 13:45:00 stopping-all geo1… geo1…
#>  7 108H         2000-01-01 12:55:00 2000-01-01 13:48:00 stopping-all geo1… geo7 
#>  8 108H         2000-01-01 12:55:00 2000-01-01 13:49:00 stopping-all geo7  geo7 
#>  9 109H         2000-01-01 13:25:00 2000-01-01 14:09:00 stopping-all geo6  geo1…
#> 10 109H         2000-01-01 13:25:00 2000-01-01 14:09:30 stopping-all geo1… geo1…
#> # ℹ 271 more rows
#> # ℹ 4 more variables: duration <Duration>, type <chr>, j <int>, T_journey <dbl>