Calculate train journey intervals

Calculates a set of time intervals containing valid Centrix observations. These observations correspond to distinct / interlaced train journeys.

Usage

calculate_time_windows(aspect_events, track_events, asset_map)

Arguments

aspect_events

A data frame containing pre-processed aspect events with columns:

signal character() signal ID.

dt lubridate::POSIXct() datetime.

aspect factor() signal aspect after train enters the signal section.

past_aspect factor() signal aspect before train enters the signal section.

track_events

A data frame containing pre-processed track events with columns:

track character() track ID.

dt lubridate::POSIXct() datetime.

occcupied logical() TRUE if train enters track, else FALSE.

event character() 'enters' if train enters track, else 'vacates'.

asset_map

A data frame containing a 1-1 mapping from signals to berths and a 1-many mapping from berths to tracks. The map also contains a 1-many mapping from tracks to events, reflecting the fact that there is an 'enter' and a 'vacate' event for each track. Therefore, the data frame should adhere to the following structure:

signal: A character() vector containing the signal ID. It should follow the regex pattern "S[0-9]+".

berth: A character() vector containing the berth ID. It should follow the regex pattern "[A-Z]+".

track: A character() vector containing the track ID. It should follow the regex pattern "T[A-Z]+(-[0-9])?".

event: A character() vector containing the event type. It must be one of either "enters" or "vacates".

Value

A data frame containing window IDs and time intervals in the columns:

window numeric() window ID.

interval lubridate::interval() time intervals.

Each window indicates a time interval in which the data contain valid observations.

Details

A train journey across the track section begins as the train enters the first track specified in the asset_map and ends when the train vacates the last. If the next train to pass through is sufficiently far behind, there will be some period of time where the track is empty - between the first train leaving the last track and the next train entering the first track. Therefore, any events that occur in this time are invalid and are discarded.

The time windows that this function identifies are the inverse of the previous intervals, defining the time periods that trains are present in the track section. The events within these windows are validated by the number of signals/tracks, the number of events, and the number of trains. Only valid journeys are returned, which can result in data loss of around 1%.

See also

Examples

data(aspect_events, track_events, asset_map)
time_windows <- calculate_time_windows(aspect_events,
                                       track_events,
                                       asset_map)
time_windows
#> # A tibble: 681 × 2
#>    window interval                                        
#>     <int> <Interval>                                      
#>  1      2 2000-01-01 13:07:11 UTC--2000-01-01 13:21:32 UTC
#>  2      3 2000-01-01 13:22:09 UTC--2000-01-01 13:26:10 UTC
#>  3      4 2000-01-01 13:27:43 UTC--2000-01-01 13:32:27 UTC
#>  4      7 2000-01-01 14:07:06 UTC--2000-01-01 14:20:32 UTC
#>  5      8 2000-01-01 14:26:30 UTC--2000-01-01 14:33:18 UTC
#>  6      9 2000-01-01 14:34:54 UTC--2000-01-01 14:54:57 UTC
#>  7     10 2000-01-01 15:06:54 UTC--2000-01-01 15:20:53 UTC
#>  8     11 2000-01-01 15:37:57 UTC--2000-01-01 15:50:21 UTC
#>  9     12 2000-01-01 16:07:15 UTC--2000-01-01 16:17:10 UTC
#> 10     13 2000-01-01 16:37:04 UTC--2000-01-01 16:49:11 UTC
#> # ℹ 671 more rows