Calculate TSAR and Sub-Components.

Calculate berth-level information including TSAR and all sub-components, e.g., T_onset, T_clear, T_offset, etc.

Usage

calculate_tsars(track_events, red_events, asset_map)

Arguments

track_events

A data frame containing windowed track events. No input validation is performed. It is assumed that the data frame will follow the structure returned by the function filter_track_events()

red_events

A data frame containing windowed red events. No input validation is performed. It is assumed that the data frame will follow the structure returns by the function filter_aspect_events()

asset_map

A data frame containing a 1-1 mapping from signals to berths and a 1-many mapping from berths to tracks. The map also contains a 1-many mapping from tracks to events, reflecting the fact that there is an 'enter' and a 'vacate' event for each track. Therefore, the data frame should adhere to the following structure:

signal: A character() vector containing the signal ID. It should follow the regex pattern "S[0-9]+".

berth: A character() vector containing the berth ID. It should follow the regex pattern "[A-Z]+".

track: A character() vector containing the track ID. It should follow the regex pattern "T[A-Z]+(-[0-9])?".

event: A character() vector containing the event type. It must be one of either "enters" or "vacates".

Value

A data frame containing berth level observations including TSAR and its sub-components. The following columns are included:

signal Signal ID.

berth Berth ID.

train_id Train ID.

aspect Signal aspect on berth entry.

t_enters Date and time of berth entry.

t_red_on Date and time of signal changing to red.

t_enters_next Date and time of entry into next berth.

t_vacates Date and time of vacating the berth.

t_red_off Date and time of signal changing from red.

TSAR Amount of time the signal spends on red.

T_onset Signal onset time between train entry and signal changing to red.

T_clear Amount of time the train takes to clear the berth.

T_offset Signal offset time between train exit and signal changing from red.

T_travel Amount of time the train takes to travel the length of the berth.

T_coach Amount of time the train takes to travel its own length.

Details

Combines valid track events from filter_track_events() with valid red signal events from filter_aspect_events() to calculate TSAR and its sub-components. The data frame returned by this function contains the first data that can be used for analysis.

Examples

data(valid_track_events, valid_aspect_events, asset_map)
valid_aspect_events
#> # A tibble: 25,834 × 6
#>    signal dt                  aspect past_aspect window
#>    <chr>  <dttm>              <fct>  <fct>        <int>
#>  1 S1     2000-01-01 13:07:12 R      G                2
#>  2 S1     2000-01-01 13:09:36 Y      R                2
#>  3 S1     2000-01-01 13:10:11 YY     Y                2
#>  4 S1     2000-01-01 13:11:00 G      YY               2
#>  5 S1     2000-01-01 13:11:59 R      G                2
#>  6 S1     2000-01-01 13:13:02 Y      R                2
#>  7 S1     2000-01-01 13:13:44 YY     Y                2
#>  8 S1     2000-01-01 13:14:35 G      YY               2
#>  9 S1     2000-01-01 13:22:10 R      G                3
#> 10 S1     2000-01-01 13:22:58 Y      R                3
#> # ℹ 25,824 more rows
#> # ℹ 1 more variable: interval <Interval>
valid_track_events
#> # A tibble: 24,332 × 6
#>    track dt                  occupied event   window
#>    <chr> <dttm>              <lgl>    <chr>    <int>
#>  1 TA-1  2000-01-01 13:07:11 TRUE     enters       2
#>  2 TA-1  2000-01-01 13:07:55 FALSE    vacates      2
#>  3 TA-1  2000-01-01 13:11:58 TRUE     enters       2
#>  4 TA-1  2000-01-01 13:12:34 FALSE    vacates      2
#>  5 TA-1  2000-01-01 13:22:09 TRUE     enters       3
#>  6 TA-1  2000-01-01 13:22:37 FALSE    vacates      3
#>  7 TA-1  2000-01-01 13:27:43 TRUE     enters       4
#>  8 TA-1  2000-01-01 13:28:14 FALSE    vacates      4
#>  9 TA-1  2000-01-01 14:07:06 TRUE     enters       7
#> 10 TA-1  2000-01-01 14:07:54 FALSE    vacates      7
#> # ℹ 24,322 more rows
#> # ℹ 1 more variable: interval <Interval>

berth_events <- calculate_tsars(valid_track_events,
                                valid_aspect_events,
                                asset_map)
berth_events
#> # A tibble: 6,952 × 15
#>    signal berth train_id aspect t_enters            t_red_on           
#>    <chr>  <chr>    <int> <fct>  <dttm>              <dttm>             
#>  1 S1     A            1 G      2000-01-01 13:07:11 2000-01-01 13:07:12
#>  2 S2     B            1 G      2000-01-01 13:09:18 2000-01-01 13:09:19
#>  3 S3     C            1 G      2000-01-01 13:09:58 2000-01-01 13:09:59
#>  4 S4     D            1 G      2000-01-01 13:10:41 2000-01-01 13:10:42
#>  5 S5     E            1 G      2000-01-01 13:12:39 2000-01-01 13:12:39
#>  6 S6     F            1 G      2000-01-01 13:13:18 2000-01-01 13:13:19
#>  7 S7     G            1 G      2000-01-01 13:15:35 2000-01-01 13:15:35
#>  8 S8     H            1 G      2000-01-01 13:16:19 2000-01-01 13:16:20
#>  9 S1     A            2 G      2000-01-01 13:11:58 2000-01-01 13:11:59
#> 10 S2     B            2 YY     2000-01-01 13:12:43 2000-01-01 13:12:43
#> # ℹ 6,942 more rows
#> # ℹ 9 more variables: t_enters_next <dttm>, t_vacates <dttm>, t_red_off <dttm>,
#> #   TSAR <dbl>, T_onset <dbl>, T_clear <dbl>, T_offset <dbl>, T_travel <dbl>,
#> #   T_coach <dbl>