Example delays

An example of a data frame containing estimated delays

Usage

delays

Format

signal

Signal ID

train_id

Train ID (Centrix)

station

Station name

T_dwell

Dwell time in seconds

arrival_delay

Arrival delay in seconds, can be negative if the train arrives early

departure_delay

Departure delay in seconds, can be negative if the train departs early

excess_dwell_time

Difference between actual dwell time and scheduled dwell time. Can be negative if the train spends less time in the station than scheduled

early

Logical vector indicating whether the train arrived at the station earlier than scheduled

Details

A data frame containing 66 row and 8 columns

Examples

delays
#> # A tibble: 66 × 8
#>    signal train_id station T_dwell arrival_delay  departure_delay
#>    <chr>     <int> <chr>     <dbl> <drtn>         <drtn>         
#>  1 S1            5 geo110    15.9  -59.52909 secs  -73.60537 secs
#>  2 S4            5 geo111    13.9  -32.17117 secs  -48.28861 secs
#>  3 S6            5 geo112    35.9  -24.92197 secs  -18.98806 secs
#>  4 S1           13 geo110    -2.13 -71.54309 secs -103.67437 secs
#>  5 S4           13 geo111    47.2  -68.05317 secs  -50.87861 secs
#>  6 S6           13 geo112    39.3  -30.00197 secs  -20.74406 secs
#>  7 S1           16 geo110     5.62 105.54191 secs   81.16663 secs
#>  8 S4           16 geo111    14.0  117.69183 secs  101.67739 secs
#>  9 S6           16 geo112    18.6  124.26103 secs  112.86294 secs
#> 10 S1           19 geo110     1.53  74.39591 secs   45.93063 secs
#> # ℹ 56 more rows
#> # ℹ 2 more variables: excess_dwell_time <drtn>, early <lgl>