Estimate berth lengths

This function estimates berth lengths based on the median travel times across each berth for unimpeded fast trains, with a given total journey time and total track length.

Usage

estimate_berth_lengths(
  timetable_specification,
  id_matching,
  berth_events_groups,
  expected_journey_time,
  track_length
)

Arguments

timetable_specification

Timetable specification

id_matching

Id matching

berth_events_groups

Berth events groups

expected_journey_time

Expected journey time in seconds from start of Centrix track to end

track_length

Total length of Centrix track in miles

Value

Berth lengths and median travel times. A data frame with 6 columns:

berthBerth ID

median_travelMedian travel time for fast trains across that berth

v.mphExpected speed for fast trains across that berth

L.milesEstimated berth length in miles

L.kmEstimated berth length in kilometres

LEstimated berth length in metres

See also

Examples

data(timetable_specification, id_matching, berth_events_groups)
timetable_specification
#> # A tibble: 281 × 10
#>    train_header dt_origin           wtt                 group        xi    xk   
#>    <chr>        <dttm>              <dttm>              <chr>        <chr> <chr>
#>  1 108H         2000-01-01 12:55:00 2000-01-01 13:39:00 stopping-all geo6  geo1…
#>  2 108H         2000-01-01 12:55:00 2000-01-01 13:39:30 stopping-all geo1… geo1…
#>  3 108H         2000-01-01 12:55:00 2000-01-01 13:42:00 stopping-all geo1… geo1…
#>  4 108H         2000-01-01 12:55:00 2000-01-01 13:42:30 stopping-all geo1… geo1…
#>  5 108H         2000-01-01 12:55:00 2000-01-01 13:44:30 stopping-all geo1… geo1…
#>  6 108H         2000-01-01 12:55:00 2000-01-01 13:45:00 stopping-all geo1… geo1…
#>  7 108H         2000-01-01 12:55:00 2000-01-01 13:48:00 stopping-all geo1… geo7 
#>  8 108H         2000-01-01 12:55:00 2000-01-01 13:49:00 stopping-all geo7  geo7 
#>  9 109H         2000-01-01 13:25:00 2000-01-01 14:09:00 stopping-all geo6  geo1…
#> 10 109H         2000-01-01 13:25:00 2000-01-01 14:09:30 stopping-all geo1… geo1…
#> # ℹ 271 more rows
#> # ℹ 4 more variables: duration <Duration>, type <chr>, j <int>, T_journey <dbl>
id_matching
#> # A tibble: 42 × 3
#>    train_id train_header dt_origin          
#>       <int> <chr>        <dttm>             
#>  1        5 109H         2000-01-01 13:25:00
#>  2        6 46H          2000-01-01 13:36:00
#>  3        9 185H         2000-01-01 13:52:00
#>  4       10 1H           2000-01-01 14:00:00
#>  5       12 16H          2000-01-01 14:15:00
#>  6       13 111H         2000-01-01 14:25:00
#>  7       14 47H          2000-01-01 14:39:00
#>  8       15 186H         2000-01-01 14:52:00
#>  9       15 2H           2000-01-01 15:00:00
#> 10       16 112H         2000-01-01 14:55:00
#> # ℹ 32 more rows
berth_events_groups
#> # A tibble: 6,952 × 16
#>    signal berth train_id aspect t_enters            t_red_on           
#>    <chr>  <chr>    <int> <fct>  <dttm>              <dttm>             
#>  1 S1     A            1 G      2000-01-01 13:07:11 2000-01-01 13:07:12
#>  2 S2     B            1 G      2000-01-01 13:09:18 2000-01-01 13:09:19
#>  3 S3     C            1 G      2000-01-01 13:09:58 2000-01-01 13:09:59
#>  4 S4     D            1 G      2000-01-01 13:10:41 2000-01-01 13:10:42
#>  5 S5     E            1 G      2000-01-01 13:12:39 2000-01-01 13:12:39
#>  6 S6     F            1 G      2000-01-01 13:13:18 2000-01-01 13:13:19
#>  7 S7     G            1 G      2000-01-01 13:15:35 2000-01-01 13:15:35
#>  8 S8     H            1 G      2000-01-01 13:16:19 2000-01-01 13:16:20
#>  9 S1     A            2 G      2000-01-01 13:11:58 2000-01-01 13:11:59
#> 10 S2     B            2 YY     2000-01-01 13:12:43 2000-01-01 13:12:43
#> # ℹ 6,942 more rows
#> # ℹ 10 more variables: t_enters_next <dttm>, t_vacates <dttm>,
#> #   t_red_off <dttm>, TSAR <dbl>, T_onset <dbl>, T_clear <dbl>, T_offset <dbl>,
#> #   T_travel <dbl>, T_coach <dbl>, group <chr>
estimate_berth_lengths(timetable_specification,
                       id_matching,
                       berth_events_groups,
                       expected_journey_time = 270,
                       track_length = 5.97)
#> # A tibble: 8 × 6
#>   berth median_travel v.mph L.miles  L.km     L
#>   <chr>         <dbl> <dbl>   <dbl> <dbl> <dbl>
#> 1 A              32.9  79.6   0.727 1.17  1170.
#> 2 B              25.1  79.6   0.555 0.893  893.
#> 3 C              27.6  79.6   0.610 0.982  982.
#> 4 D              24.7  79.6   0.546 0.879  879.
#> 5 E              24.7  79.6   0.546 0.879  879.
#> 6 F              28.9  79.6   0.639 1.03  1028.
#> 7 G              28.9  79.6   0.638 1.03  1027.
#> 8 H              24.2  79.6   0.536 0.863  863.