Calculate delays from Centrix and timetable data

Calculate delays from estimated dwell times and timetable data. This function calculates arrival delay, departure delay, excess dwell time, and whether a train was early or late. These are all estimates, as the dwell times on which they are based are also estimates.

Usage

estimate_delays(dwell_times, timetable_subset, id_matching)

Arguments

dwell_times

Data frame containing estimated dwell times. The structure is expected to match the output of estimate_dwell_times(). However, only 6 columns are required for this function to work:

signalSignal ID

train_idTrain ID (Centrix)

stationStation name

t_entersDate and time that train enters berth

T_move_oneMoving time before coming to a stop at the station

T_dwellDwell time at the station

timetable_subset

Processed timetable containing calling patterns. The structure is expected to match the output of wrangle_timetable(). However, only 7 columns are required:

train_headerTrain ID (timetable)

dt_originDate and time of train origin

geoName of location

groupGroup name

eventEvent name; anything other than 'Arrive' or 'Depart' is ignored

wttScheduled time of event

tActual time of event

id_matching

Data frame containing a 1-1 mapping from Centrix IDs to timetable IDs. There should be 3 columns:

train_idTrain ID (Centrix)

train_headerTrain ID (timetable)

dt_originDate and time of train origin

Value

A data frame containing 8 columns:

signalSignal ID

train_idTrain ID (Centrix)

stationStation name

T_dwellDwell time in seconds

arrival_delayArrival delay in seconds, can be negative if the train arrives early

departure_delayDeparture delay in seconds, can be negative if the train departs early

excess_dwell_timeDifference between actual dwell time and scheduled dwell time. Can be negative if the train spends less time in the station than scheduled

earlyLogical vector indicating whether the train arrived at the station earlier than scheduled

See also

Examples

data(dwell_times, timetable_subset, id_matching)
dwell_times
#> # A tibble: 1,581 × 22
#>    signal berth station train_id group      t_enters            T_travel v_entry
#>    <chr>  <chr> <chr>      <int> <chr>      <dttm>                 <dbl>   <dbl>
#>  1 S1     A     geo110         1 stopping-… 2000-01-01 13:07:11     127.    23.5
#>  2 S4     D     geo111         1 stopping-… 2000-01-01 13:10:41     118.    22.6
#>  3 S6     F     geo112         1 stopping-… 2000-01-01 13:13:18     137.    22.4
#>  4 S1     A     geo110         5 stopping-… 2000-01-01 14:07:06     128.    23.5
#>  5 S4     D     geo111         5 stopping-… 2000-01-01 14:10:40     111.    21.4
#>  6 S6     F     geo112         5 stopping-… 2000-01-01 14:13:14     141.    20.6
#>  7 S1     A     geo110        11 stopping-… 2000-01-01 14:40:42     125.    23.5
#>  8 S4     D     geo111        11 stopping-… 2000-01-01 14:44:20     111.    19.6
#>  9 S6     F     geo112        11 stopping-… 2000-01-01 14:46:53     127.    20.6
#> 10 S1     A     geo110        13 stopping-… 2000-01-01 15:06:54     110.    23.5
#> # ℹ 1,571 more rows
#> # ℹ 14 more variables: a_brake <dbl>, L1 <dbl>, v_exit <dbl>, a_tract <dbl>,
#> #   L2 <dbl>, T_brake <dbl>, S_brake <dbl>, T_const_one <dbl>,
#> #   T_move_one <dbl>, T_tract <dbl>, S_tract <dbl>, T_const_two <dbl>,
#> #   T_move_two <dbl>, T_dwell <dbl>
timetable_subset
#> # A tibble: 792 × 12
#>    train_header dt_origin           geo    event  wtt                
#>    <chr>        <dttm>              <chr>  <chr>  <dttm>             
#>  1 108H         2000-01-01 12:55:00 geo5   Arrive 2000-01-01 13:34:00
#>  2 108H         2000-01-01 12:55:00 geo5   Depart 2000-01-01 13:35:30
#>  3 108H         2000-01-01 12:55:00 geo6   Pass   2000-01-01 13:37:30
#>  4 108H         2000-01-01 12:55:00 geo110 Arrive 2000-01-01 13:39:00
#>  5 108H         2000-01-01 12:55:00 geo110 Depart 2000-01-01 13:39:30
#>  6 108H         2000-01-01 12:55:00 geo111 Arrive 2000-01-01 13:42:00
#>  7 108H         2000-01-01 12:55:00 geo111 Depart 2000-01-01 13:42:30
#>  8 108H         2000-01-01 12:55:00 geo112 Arrive 2000-01-01 13:44:30
#>  9 108H         2000-01-01 12:55:00 geo112 Depart 2000-01-01 13:45:00
#> 10 108H         2000-01-01 12:55:00 geo7   Arrive 2000-01-01 13:48:00
#> # ℹ 782 more rows
#> # ℹ 7 more variables: t <dttm>, delay <dbl>, allow <dbl>, allow_perf <dbl>,
#> #   allow_path <dbl>, allow_eng <dbl>, group <chr>
id_matching
#> # A tibble: 42 × 3
#>    train_id train_header dt_origin          
#>       <int> <chr>        <dttm>             
#>  1        5 109H         2000-01-01 13:25:00
#>  2        6 46H          2000-01-01 13:36:00
#>  3        9 185H         2000-01-01 13:52:00
#>  4       10 1H           2000-01-01 14:00:00
#>  5       12 16H          2000-01-01 14:15:00
#>  6       13 111H         2000-01-01 14:25:00
#>  7       14 47H          2000-01-01 14:39:00
#>  8       15 186H         2000-01-01 14:52:00
#>  9       15 2H           2000-01-01 15:00:00
#> 10       16 112H         2000-01-01 14:55:00
#> # ℹ 32 more rows
delays <- estimate_delays(dwell_times, timetable_subset, id_matching)
delays
#> # A tibble: 66 × 8
#>    signal train_id station T_dwell arrival_delay  departure_delay
#>    <chr>     <int> <chr>     <dbl> <drtn>         <drtn>         
#>  1 S1            5 geo110    15.9  -59.52909 secs  -73.60537 secs
#>  2 S4            5 geo111    13.9  -32.17117 secs  -48.28861 secs
#>  3 S6            5 geo112    35.9  -24.92197 secs  -18.98806 secs
#>  4 S1           13 geo110    -2.13 -71.54309 secs -103.67437 secs
#>  5 S4           13 geo111    47.2  -68.05317 secs  -50.87861 secs
#>  6 S6           13 geo112    39.3  -30.00197 secs  -20.74406 secs
#>  7 S1           16 geo110     5.62 105.54191 secs   81.16663 secs
#>  8 S4           16 geo111    14.0  117.69183 secs  101.67739 secs
#>  9 S6           16 geo112    18.6  124.26103 secs  112.86294 secs
#> 10 S1           19 geo110     1.53  74.39591 secs   45.93063 secs
#> # ℹ 56 more rows
#> # ℹ 2 more variables: excess_dwell_time <drtn>, early <lgl>