Estimate dwell times

This function estimates dwell times from Centrix data

Usage

estimate_dwell_times(
  berth_events_groups,
  berth_lengths,
  station_names,
  stopping_patterns,
  a_brake,
  a_tract,
  station_berth_lengths = NULL
)

Arguments

berth_events_groups

Berth-level Centrix data

berth_lengths

Data frame with 2 columns:

berthBerth ID

LBerth length in metres

station_names

Data frame with 2 columns:

berthBerth ID

stationStation name

This provides a 1-1 mapping from berth ID to station name. The station names should correspond to those in the timetable

stopping_patterns

Data frame with 2 columns:

groupGroup name

stationStation name

This provides a list of stations that each group stops at.

a_brake

Braking capacity in m/s^2

a_tract

Acceleration in m/s^2

station_berth_lengths

(Optional) Data frame with 3 columns:

stationStation name

L1Distance from start of berth to station

L2Distance from station to end of berth

If station_berth_lengths is NULL (the default), then L1 and L2 will be equal to the berth length halved

Value

Data frame with 22 columns:

signalSignal ID

berthBerth ID

stationStation name

train_idTrain ID

groupGroup name

t_entersDate and time that train enters berth

T_travelBerth travel time

v_entryVelocity at entry to berth in m/s

a_brakeBraking capacity in m/s^2

L1Distance from start of berth to station

v_exitVelocity at exit of berth in m/s

a_tractAcceleration in m/s^2

L2Distance from station to end of berth

T_brakeTime spent braking

S_brakeDistance spent braking

T_const_oneTime spent moving at a constant speed before braking for station

T_move_oneTotal time moving before reaching station

T_tractTime spent accelerating

S_tractDistance spent accelerating

T_const_twoTime spent moving at a constant speed before vacating berth

T_move_twoTotal time moving after leaving station

T_dwellDwell time at station

Examples

data(berth_events_groups, berth_lengths)

# Set up station mapping
station_names <- dplyr::tribble(
  ~berth, ~station,
  "A", "geo110",
  "D", "geo111",
  "F", "geo112"
)

# Specify stopping patterns
stopping_patterns <- dplyr::tribble(
  ~group, ~station,
  "stopping-all", "geo110",
  "stopping-all", "geo111",
  "stopping-all", "geo112",
  "stopping-geo112", "geo112"
)

# Define braking and accelerating capacities
a_brake = 0.4
a_tract = 0.35

# Calculate dwell times
dwell_times <- estimate_dwell_times(berth_events_groups,
                                    berth_lengths,
                                    station_names,
                                    stopping_patterns,
                                    a_brake,
                                    a_tract)

# View dwell times
dwell_times
#> # A tibble: 1,581 × 22
#>    signal berth station train_id group      t_enters            T_travel v_entry
#>    <chr>  <chr> <chr>      <int> <chr>      <dttm>                 <dbl>   <dbl>
#>  1 S1     A     geo110         1 stopping-… 2000-01-01 13:07:11     127.    23.5
#>  2 S4     D     geo111         1 stopping-… 2000-01-01 13:10:41     118.    22.6
#>  3 S6     F     geo112         1 stopping-… 2000-01-01 13:13:18     137.    22.4
#>  4 S1     A     geo110         5 stopping-… 2000-01-01 14:07:06     128.    23.5
#>  5 S4     D     geo111         5 stopping-… 2000-01-01 14:10:40     111.    21.4
#>  6 S6     F     geo112         5 stopping-… 2000-01-01 14:13:14     141.    20.6
#>  7 S1     A     geo110        11 stopping-… 2000-01-01 14:40:42     125.    23.5
#>  8 S4     D     geo111        11 stopping-… 2000-01-01 14:44:20     111.    19.6
#>  9 S6     F     geo112        11 stopping-… 2000-01-01 14:46:53     127.    20.6
#> 10 S1     A     geo110        13 stopping-… 2000-01-01 15:06:54     110.    23.5
#> # ℹ 1,571 more rows
#> # ℹ 14 more variables: a_brake <dbl>, L1 <dbl>, v_exit <dbl>, a_tract <dbl>,
#> #   L2 <dbl>, T_brake <dbl>, S_brake <dbl>, T_const_one <dbl>,
#> #   T_move_one <dbl>, T_tract <dbl>, S_tract <dbl>, T_const_two <dbl>,
#> #   T_move_two <dbl>, T_dwell <dbl>

# Summarise dwell times
# There are some negative values, which indicate outliers and potentially invalid observations
summary(dwell_times$T_dwell)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>  -50.25   14.82   21.67   29.08   33.13  113.56