Filter Centrix events by time intervals

Filter Centrix events to only those that occur within the calculated valid time intervals.

filter_track_events() filters to only the tracks specified in the asset map.

filter_aspect_events() filters to only the signals specified in the asset map.

Usage

filter_aspect_events(aspect_events, time_windows, asset_map)

filter_track_events(track_events, time_windows, asset_map)

Arguments

aspect_events

A data frame containing pre-processed aspect events with columns:

signal character() signal ID.

dt lubridate::POSIXct() datetime.

aspect factor() signal aspect after train enters the signal section.

past_aspect factor() signal aspect before train enters the signal section.

time_windows

A data frame containing window IDs and time intervals in the columns:

window numeric() window ID.

interval lubridate::interval() time intervals.

Each window indicates a time interval in which the data contain valid observations.

asset_map

A data frame containing a 1-1 mapping from signals to berths and a 1-many mapping from berths to tracks. The map also contains a 1-many mapping from tracks to events, reflecting the fact that there is an 'enter' and a 'vacate' event for each track. Therefore, the data frame should adhere to the following structure:

signal: A character() vector containing the signal ID. It should follow the regex pattern "S[0-9]+".

berth: A character() vector containing the berth ID. It should follow the regex pattern "[A-Z]+".

track: A character() vector containing the track ID. It should follow the regex pattern "T[A-Z]+(-[0-9])?".

event: A character() vector containing the event type. It must be one of either "enters" or "vacates".

track_events

A data frame containing pre-processed track events with columns:

track character() track ID.

dt lubridate::POSIXct() datetime.

occcupied logical() TRUE if train enters track, else FALSE.

event character() 'enters' if train enters track, else 'vacates'.

Value

A data frame containing filtered events, with additional columns:

window numeric() the window ID.

interval lubridate::interval() the time interval.

See also

Examples

data(aspect_events, time_windows, asset_map)
aspect_events
#> # A tibble: 32,512 × 4
#>    signal dt                  aspect past_aspect
#>    <chr>  <dttm>              <fct>  <fct>      
#>  1 S1     2000-01-01 06:13:49 R      NA         
#>  2 S1     2000-01-01 06:25:48 YY     R          
#>  3 S1     2000-01-01 06:25:48 Y      YY         
#>  4 S1     2000-01-01 06:25:53 R      Y          
#>  5 S1     2000-01-01 06:26:16 YY     R          
#>  6 S1     2000-01-01 06:26:16 Y      YY         
#>  7 S1     2000-01-01 07:44:16 R      Y          
#>  8 S1     2000-01-01 08:06:31 YY     R          
#>  9 S1     2000-01-01 08:06:31 Y      YY         
#> 10 S1     2000-01-01 08:20:08 R      Y          
#> # ℹ 32,502 more rows
valid_aspect_events <- filter_aspect_events(aspect_events,
                                            time_windows,
                                            asset_map)
valid_aspect_events
#> # A tibble: 25,834 × 6
#>    signal dt                  aspect past_aspect window
#>    <chr>  <dttm>              <fct>  <fct>        <int>
#>  1 S1     2000-01-01 13:07:12 R      G                2
#>  2 S1     2000-01-01 13:09:36 Y      R                2
#>  3 S1     2000-01-01 13:10:11 YY     Y                2
#>  4 S1     2000-01-01 13:11:00 G      YY               2
#>  5 S1     2000-01-01 13:11:59 R      G                2
#>  6 S1     2000-01-01 13:13:02 Y      R                2
#>  7 S1     2000-01-01 13:13:44 YY     Y                2
#>  8 S1     2000-01-01 13:14:35 G      YY               2
#>  9 S1     2000-01-01 13:22:10 R      G                3
#> 10 S1     2000-01-01 13:22:58 Y      R                3
#> # ℹ 25,824 more rows
#> # ℹ 1 more variable: interval <Interval>

#----------------------------------------------------------------------------

data(track_events)
track_events
#> # A tibble: 25,242 × 4
#>    track dt                  occupied event  
#>    <chr> <dttm>              <lgl>    <chr>  
#>  1 TA-1  2000-01-01 06:13:48 TRUE     enters 
#>  2 TA-1  2000-01-01 06:25:29 FALSE    vacates
#>  3 TA-1  2000-01-01 06:25:31 TRUE     enters 
#>  4 TA-1  2000-01-01 06:25:43 FALSE    vacates
#>  5 TA-1  2000-01-01 06:25:52 TRUE     enters 
#>  6 TA-1  2000-01-01 06:25:54 FALSE    vacates
#>  7 TA-1  2000-01-01 06:25:57 TRUE     enters 
#>  8 TA-1  2000-01-01 06:25:59 FALSE    vacates
#>  9 TA-1  2000-01-01 06:26:02 TRUE     enters 
#> 10 TA-1  2000-01-01 06:26:11 FALSE    vacates
#> # ℹ 25,232 more rows
valid_track_events <- filter_track_events(track_events,
                                          time_windows,
                                          asset_map)
valid_track_events
#> # A tibble: 24,332 × 6
#>    track dt                  occupied event   window
#>    <chr> <dttm>              <lgl>    <chr>    <int>
#>  1 TA-1  2000-01-01 13:07:11 TRUE     enters       2
#>  2 TA-1  2000-01-01 13:07:55 FALSE    vacates      2
#>  3 TA-1  2000-01-01 13:11:58 TRUE     enters       2
#>  4 TA-1  2000-01-01 13:12:34 FALSE    vacates      2
#>  5 TA-1  2000-01-01 13:22:09 TRUE     enters       3
#>  6 TA-1  2000-01-01 13:22:37 FALSE    vacates      3
#>  7 TA-1  2000-01-01 13:27:43 TRUE     enters       4
#>  8 TA-1  2000-01-01 13:28:14 FALSE    vacates      4
#>  9 TA-1  2000-01-01 14:07:06 TRUE     enters       7
#> 10 TA-1  2000-01-01 14:07:54 FALSE    vacates      7
#> # ℹ 24,322 more rows
#> # ℹ 1 more variable: interval <Interval>