Filter timetable to relevant trains

filter_relevant_services() filters only by station, ignoring direction.

filter_relevant_direction() filters by both station and direction. It is assumed that stations is in the forward direction, i.e., the first element is at the start of the track section and the last element is at the end.

Usage

filter_relevant_services(timetable, stations)

filter_relevant_direction(timetable, stations)

Arguments

timetable

Data frame containing timetable data. Strict input validation is performed. Must be a data frame with the following columns:

train_headercharacter() Train identifier

dt_originlubridate::POSIXct() Datetime at which the train originated

geocharacter() Name of event location

eventcharacter() Type of event, must only contain the following values: Pass, Arrive, Depart, Originate, Terminate

wttlubridate::POSIXct() Scheduled datetime of event

tlubridate::POSIXct() Actual datetime of event

delaynumeric() Difference between actual time and scheduled time of event

allownumeric() Timetabled delay allowance

stations

A list of TIPLOCs (geo) through which trains pass. This list is used to work out which trains travel through the track section of interest. For example, there could be a fork in the track, where trains can travel to the left or to the right. If you want to only look at trains which turn to the left, stations would include all the TIPLOCs of interest which are situated along the left track. Any trains which do not pass through any of the TIPLOCs specified in stations will be discarded.

Value

filter_relevant_services returns a (usually) smaller timetable only containing trains which pass through or stop at the specified stations.

filter_relevant_direction returns a (usually) smaller timetable only containing trains which pass through or stop at the specified stations in the given order.

See also

Examples

data(timetable, stations)
timetable
#> # A tibble: 8,958 × 11
#>    train_header dt_origin           geo    event     wtt                
#>    <chr>        <dttm>              <chr>  <chr>     <dttm>             
#>  1 168H         2000-01-01 12:05:00 geo1   Originate 2000-01-01 12:20:00
#>  2 168H         2000-01-01 12:05:00 geo18  Arrive    2000-01-01 12:23:00
#>  3 168H         2000-01-01 12:05:00 geo18  Depart    2000-01-01 12:24:00
#>  4 168H         2000-01-01 12:05:00 geo2   Arrive    2000-01-01 12:28:00
#>  5 168H         2000-01-01 12:05:00 geo2   Depart    2000-01-01 12:29:00
#>  6 168H         2000-01-01 12:05:00 geo3   Arrive    2000-01-01 12:35:30
#>  7 168H         2000-01-01 12:05:00 geo3   Depart    2000-01-01 12:36:30
#>  8 168H         2000-01-01 12:05:00 geo104 Arrive    2000-01-01 12:39:00
#>  9 168H         2000-01-01 12:05:00 geo104 Depart    2000-01-01 12:39:30
#> 10 168H         2000-01-01 12:05:00 geo4   Arrive    2000-01-01 12:41:30
#> # ℹ 8,948 more rows
#> # ℹ 6 more variables: t <dttm>, delay <dbl>, allow <dbl>, allow_perf <dbl>,
#> #   allow_path <dbl>, allow_eng <dbl>
timetable_services <- filter_relevant_services(timetable, stations)
timetable_services
#> # A tibble: 881 × 11
#>    train_header dt_origin           geo    event  wtt                
#>    <chr>        <dttm>              <chr>  <chr>  <dttm>             
#>  1 151H         2000-01-01 12:42:00 geo5   Arrive 2000-01-01 13:25:00
#>  2 151H         2000-01-01 12:42:00 geo5   Depart 2000-01-01 13:26:00
#>  3 108H         2000-01-01 12:55:00 geo5   Arrive 2000-01-01 13:34:00
#>  4 108H         2000-01-01 12:55:00 geo5   Depart 2000-01-01 13:35:30
#>  5 108H         2000-01-01 12:55:00 geo6   Pass   2000-01-01 13:37:30
#>  6 108H         2000-01-01 12:55:00 geo110 Arrive 2000-01-01 13:39:00
#>  7 108H         2000-01-01 12:55:00 geo110 Depart 2000-01-01 13:39:30
#>  8 108H         2000-01-01 12:55:00 geo111 Arrive 2000-01-01 13:42:00
#>  9 108H         2000-01-01 12:55:00 geo111 Depart 2000-01-01 13:42:30
#> 10 108H         2000-01-01 12:55:00 geo112 Arrive 2000-01-01 13:44:30
#> # ℹ 871 more rows
#> # ℹ 6 more variables: t <dttm>, delay <dbl>, allow <dbl>, allow_perf <dbl>,
#> #   allow_path <dbl>, allow_eng <dbl>
timetable_direction <- filter_relevant_direction(timetable, stations)
timetable_direction
#> # A tibble: 792 × 11
#> # Groups:   train_header, dt_origin [123]
#>    train_header dt_origin           geo    event  wtt                
#>    <chr>        <dttm>              <chr>  <chr>  <dttm>             
#>  1 108H         2000-01-01 12:55:00 geo5   Arrive 2000-01-01 13:34:00
#>  2 108H         2000-01-01 12:55:00 geo5   Depart 2000-01-01 13:35:30
#>  3 108H         2000-01-01 12:55:00 geo6   Pass   2000-01-01 13:37:30
#>  4 108H         2000-01-01 12:55:00 geo110 Arrive 2000-01-01 13:39:00
#>  5 108H         2000-01-01 12:55:00 geo110 Depart 2000-01-01 13:39:30
#>  6 108H         2000-01-01 12:55:00 geo111 Arrive 2000-01-01 13:42:00
#>  7 108H         2000-01-01 12:55:00 geo111 Depart 2000-01-01 13:42:30
#>  8 108H         2000-01-01 12:55:00 geo112 Arrive 2000-01-01 13:44:30
#>  9 108H         2000-01-01 12:55:00 geo112 Depart 2000-01-01 13:45:00
#> 10 108H         2000-01-01 12:55:00 geo7   Arrive 2000-01-01 13:48:00
#> # ℹ 782 more rows
#> # ℹ 6 more variables: t <dttm>, delay <dbl>, allow <dbl>, allow_perf <dbl>,
#> #   allow_path <dbl>, allow_eng <dbl>