Find calling patterns from timetable

Find calling patterns from timetable

Usage

find_calling_patterns(timetable, stopping_stations)

Arguments

timetable

Data frame containing timetable data. Strict input validation is performed. Must be a data frame with the following columns:

train_headercharacter() Train identifier

dt_originlubridate::POSIXct() Datetime at which the train originated

geocharacter() Name of event location

eventcharacter() Type of event, must only contain the following values: Pass, Arrive, Depart, Originate, Terminate

wttlubridate::POSIXct() Scheduled datetime of event

tlubridate::POSIXct() Actual datetime of event

delaynumeric() Difference between actual time and scheduled time of event

allownumeric() Timetabled delay allowance

stopping_stations

A subset of stations that specifies which stations trains can stop at. This is used for calculating calling patterns. Therefore, any train which does not stop at any of the stations specified in stopping_stations will be labeled as a 'fast' train. Trains which do stop at these stations will be labeled with 'stopping-', with the specific stations at which they stop attached as suffixes. For example, if a train stops at 'geo10', its calling pattern will be labeled as 'stopping-geo10'.

Value

The timetable with calling patterns added.

See also

Examples

data(timetable, stations, stopping_stations)
calling_patterns <- timetable |>
                      filter_relevant_direction(stations) |>
                      find_calling_patterns(stopping_stations)
calling_patterns[,c("train_header", "dt_origin", "geo", "event", "group")]
#> # A tibble: 792 × 5
#>    train_header dt_origin           geo    event  group       
#>    <chr>        <dttm>              <chr>  <chr>  <chr>       
#>  1 108H         2000-01-01 12:55:00 geo5   Arrive stopping-all
#>  2 108H         2000-01-01 12:55:00 geo5   Depart stopping-all
#>  3 108H         2000-01-01 12:55:00 geo6   Pass   stopping-all
#>  4 108H         2000-01-01 12:55:00 geo110 Arrive stopping-all
#>  5 108H         2000-01-01 12:55:00 geo110 Depart stopping-all
#>  6 108H         2000-01-01 12:55:00 geo111 Arrive stopping-all
#>  7 108H         2000-01-01 12:55:00 geo111 Depart stopping-all
#>  8 108H         2000-01-01 12:55:00 geo112 Arrive stopping-all
#>  9 108H         2000-01-01 12:55:00 geo112 Depart stopping-all
#> 10 108H         2000-01-01 12:55:00 geo7   Arrive stopping-all
#> # ℹ 782 more rows