Match Centrix and Timetable IDs

match_ids() matches Centrix observations with timetabling information. The result is a data set that contains a 1-1 mapping between Centrix IDs and timetable IDs. This is a necessary step because the timings used in the timetable are rounded to the nearest 30 seconds, resulting in much lower granularity. There may also be other inaccuracies with the timetable. This function results in data loss, the amount of which depends on how accurate the matching needs to be.

Usage

match_ids(berth_events_groups, timetable_groups, match_mapping)

Arguments

berth_events_groups

Data frame containing berth-level Centrix data. The structure is expected to conform to the following:

berthcharacter() Berth ID

train_idnumeric() Train ID

t_enterslubridate::POSIXct() Date and time that the train enters the berth

t_vacateslubridate::POSIXct() Date and time that the train vacates the berth

groupcharacter() Group name, corresponding to the train's calling pattern

timetable_groups

Data frame containing Timetable data. The structure is expected to conform to the following:

train_headercharacter() Train ID

dt_originlubridate::POSIXct() Date and time that the service originated

groupcharacter() Group name, corresponding to the train's calling pattern

geocharacter() Name of TIPLOCs that trains pass through or stop at

eventcharacter() Type of timetable event, every element must be one of 'Arrive', 'Depart', or 'Pass'

wttlubridate::POSIXct() Scheduled date and time of event

tlubridate::POSIXct() Actual date and time of event

match_mapping

Data frame containing instructions for the matching. It should match the following structure:

groupcharacter() Group name, corresponding to the train's calling pattern

berthcharacter() Berth ID

geocharacter() Name of TIPLOC(s) or station(s) to match at

lbnumeric() Number of seconds to shift the start of the matching window

ubnumeric() Number of seconds to shift the end of the matching window

Value

Data frame containing a 1-1 mapping between Centrix IDs and Timetable IDs.

Details

There are two methods for matching Centrix data with timetable data. The method that is used depends on whether the train is a fast train (does not stop at any stations present in the Centrix data set) or not.

For fast trains, matches must be found at the nearest TIPLOC, e.g., a junction preceding the Centrix track section. For these trains, the lb and ub columns of the match_mapping control what is considered a match. If the train passed through the junction at time t_Pass and the train entered the Centrix track at time t_enters, a match is made if:
t_Pass > t_enters + lb & t_Pass < t_enters + ub
I.e., a match is made if the train passed through the junction less than lb seconds before t_enters (assuming that lb is negative) and if the train passed through the junction earlier than t_enters + ub. Thus, lb and ub define a time window based on t_enters in which t_Pass must occur for a match to be made.

For stopping trains, matches can be made at stations, which is conceptually simpler. Let's take the example of a train which stops at only one station. The train enters the station's berth at time t_enters, arrives at the station at t_Arrive, departs the station at t_Depart and vacates the berth at time t_vacates. A match is made if:
t_enters + lb < t_Arrive & t_Depart < t_vacates + ub
I.e., a match is made if the train arrives at the station after it enters the berth and if the train departs the station before it vacates the berth. Similar to the fast trains, t_enters and t_vacates define a time window in which t_Arrive and t_Depart must occur for a match to be made. This window can be adjusted using lb and ub.

See also

Examples

# Define the matching instructions
data(berth_events_groups, timetable_subset)
match_mapping <- dplyr::tribble(
  ~group, ~berth, ~geo, ~lb, ~ub,
  # match fast trains at berth "A" and geo "geo6"
  # move start of window 240 seconds before `t_enters`
  # move end of window 60 seconds after `t_enters`
  "fast", "A", "geo6", -240, 60,
  # match stopping-all trains at berths "A","D","F", and geos "geo110","geo111","geo112"
  # don't adjust window
  "stopping-all", "A", "geo110", 0, 0,
  "stopping-all", "D", "geo111", 0, 0,
  "stopping-all", "F", "geo112", 0, 0
)

timetable_groups <- dplyr::filter(timetable_subset,
                                  event %in% c("Arrive", "Depart", "Pass"))

matched_ids <- match_ids(berth_events_groups, timetable_groups, match_mapping)
matched_ids
#> # A tibble: 42 × 3
#>    train_id train_header dt_origin          
#>       <int> <chr>        <dttm>             
#>  1        5 109H         2000-01-01 13:25:00
#>  2        6 46H          2000-01-01 13:36:00
#>  3        9 185H         2000-01-01 13:52:00
#>  4       10 1H           2000-01-01 14:00:00
#>  5       12 16H          2000-01-01 14:15:00
#>  6       13 111H         2000-01-01 14:25:00
#>  7       14 47H          2000-01-01 14:39:00
#>  8       15 186H         2000-01-01 14:52:00
#>  9       15 2H           2000-01-01 15:00:00
#> 10       16 112H         2000-01-01 14:55:00
#> # ℹ 32 more rows

match_mapping <- dplyr::tribble(
  ~group, ~berth, ~geo, ~lb, ~ub,
  "fast", "A", "geo6", -240, 60,
  # move start of window for stopping-all trains to 60 seconds before `t_enters`
  "stopping-all", "A", "geo110", -60, 0,
  "stopping-all", "D", "geo111", -60, 0,
  "stopping-all", "F", "geo112", -60, 0
)
wider_window <- match_ids(berth_events_groups, timetable_groups, match_mapping)
# More matches are found with the wider window
wider_window
#> # A tibble: 45 × 3
#>    train_id train_header dt_origin          
#>       <int> <chr>        <dttm>             
#>  1        5 109H         2000-01-01 13:25:00
#>  2        6 46H          2000-01-01 13:36:00
#>  3        9 185H         2000-01-01 13:52:00
#>  4       10 1H           2000-01-01 14:00:00
#>  5       11 110H         2000-01-01 13:55:00
#>  6       12 16H          2000-01-01 14:15:00
#>  7       13 111H         2000-01-01 14:25:00
#>  8       14 47H          2000-01-01 14:39:00
#>  9       15 186H         2000-01-01 14:52:00
#> 10       15 2H           2000-01-01 15:00:00
#> # ℹ 35 more rows