Preprocess raw Centrix data

preprocess_signal_events() handles only signals, defined by the asset string starting with the character 'S'.

preprocess_track_events() handles only tracks, defined by the asset string starting with the character 'T'.

Usage

preprocess_signal_events(raw_centrix, asset_map, state_map)

preprocess_track_events(raw_centrix, asset_map)

Arguments

raw_centrix

A data frame containing raw Centrix data. Strict input validation is applied to ensure the data adhere to the expected structure. The data should be in a data frame with the following columns:

asset: A character vector containing the asset ID. Tracks are expected to follow the regex pattern "T[A-Z]+(-[0-9])?". E.g., 'TABC' 'TABC-1' 'TABC-2'. Signals are expected to follow the regex pattern "S[0-9]+\s[A-Z]+". E.g., 'S123 HGE' 'S123 I'. The second part of the signal ID should be a state code referring to the signal aspect, e.g. 'RGE' for red or 'DGE' for green. For other codes see: https://wiki.openraildata.com/index.php/Signalling_Nomenclature#Signals.

dt: A lubridate::POSIXct object representing the date and time at which the observation was made.

transition: A character vector representing the state transition that caused the observation. A transition can be one of "UP to DN" or "DN to UP", anything else will be ignored. For tracks, "UP to DN" signifies a train entering the track, while "DN to UP" is a train vacating the track. For signals, "UP to DN" represents an aspect change. Therefore, any signals with a "DN to UP" transition will be ignored.

asset_map

A data frame containing a 1-1 mapping from signals to berths and a 1-many mapping from berths to tracks. The map also contains a 1-many mapping from tracks to events, reflecting the fact that there is an 'enter' and a 'vacate' event for each track. Therefore, the data frame should adhere to the following structure:

signal: A character vector containing the signal ID. It should follow the regex pattern "S[0-9]+".

berth: A character vector containing the berth ID. It should follow the regex pattern "[A-Z]+".

track: A character vector containing the track ID. It should follow the regex pattern "T[A-Z]+(-[0-9])?".

event: A character vector containing the event type. It must be one of either "enters" or "vacates".

state_map

A data frame containing a 1-1 mapping from signal state to signal aspect. See state_mapping (the default) for the expected structure.

Value

preprocess_signal_events returns a data frame containing aspect events with columns:

signal (character) signal ID.

dt (lubridate::POSIXct) datetime.

aspect (factor) signal aspect after train enters the signal section.

past_aspect (factor) signal aspect before train enters the signal section.

preprocess_track_events returns a data frame containing track events with columns:

track (character) track ID.

dt (lubridate::POSIXct) datetime.

occcupied (logical) TRUE if train enters track, else FALSE.

event (character) 'enters' if train enters track, else 'vacates'.

Details

preprocess_signal_events() converts the raw asset string into signal ID and state, furthering processing the state into a signal aspect.

preprocess_track_events() converts the transition data into track entry and exit events.

See also