Wrangle Raw Centrix Data

This function takes in raw Centrix data as an input and returns a fully processed data frame containing berth-level signal and track observations, along with calculations of TSAR and all its subcomponents.

Usage

wrangle_centrix(raw_centrix, asset_map, state_map = state_mapping)

Arguments

raw_centrix

A data frame containing raw Centrix data. Strict input validation is applied to ensure the data adhere to the expected structure. The data should be in a data frame with the following columns:

asset: A character() vector containing the asset ID. Tracks are expected to follow the regex pattern "T[A-Z]+(-[0-9])?". E.g., 'TABC' 'TABC-1' 'TABC-2'. Signals are expected to follow the regex pattern "S[0-9]+\s[A-Z]+". E.g., 'S123 HGE' 'S123 I'. The second part of the signal ID should be a state code referring to the signal aspect, e.g. 'RGE' for red or 'DGE' for green. For other codes see: https://wiki.openraildata.com/index.php/Signalling_Nomenclature#Signals.

dt: A lubridate::POSIXct() object representing the date and time at which the observation was made.

transition: A character() vector representing the state transition that caused the observation. A transition can be one of "UP to DN" or "DN to UP", anything else will be ignored. For tracks, "UP to DN" signifies a train entering the track, while "DN to UP" is a train vacating the track. For signals, "UP to DN" represents an aspect change. Therefore, any signals with a "DN to UP" transition will be ignored.

asset_map

A data frame containing a 1-1 mapping from signals to berths and a 1-many mapping from berths to tracks. The map also contains a 1-many mapping from tracks to events, reflecting the fact that there is an 'enter' and a 'vacate' event for each track. Therefore, the data frame should adhere to the following structure:

signal: A character() vector containing the signal ID. It should follow the regex pattern "S[0-9]+".

berth: A character() vector containing the berth ID. It should follow the regex pattern "[A-Z]+".

track: A character() vector containing the track ID. It should follow the regex pattern "T[A-Z]+(-[0-9])?".

event: A character() vector containing the event type. It must be one of either "enters" or "vacates".

state_map

A data frame containing a 1-1 mapping from signal state to signal aspect. See state_mapping (the default) for the expected structure.

Value

A data frame containing the fully processed Centrix data, containing berth-level information about signal and track events. This includes TSARs and all sub-components, e.g. 'T_travel' and 'T_offset'.

Details

This function wraps:

preprocess_signal_events() preprocess_track_events()

calculate_time_windows()

filter_aspect_events() filter_track_events()

calculate_tsars()

Examples

data(raw_centrix, asset_map)
raw_centrix
#> # A tibble: 110,339 × 3
#>    asset   dt                  transition
#>    <chr>   <dttm>              <chr>     
#>  1 TA-1 TR 2000-01-01 06:13:48 UP to DN  
#>  2 S1 HHGE 2000-01-01 06:13:49 UP to DN  
#>  3 S1 HGE  2000-01-01 06:13:49 UP to DN  
#>  4 S1 RGE  2000-01-01 06:13:49 DN to UP  
#>  5 TA-1 TR 2000-01-01 06:25:29 DN to UP  
#>  6 TA-1 TR 2000-01-01 06:25:31 UP to DN  
#>  7 TA-1 TR 2000-01-01 06:25:43 DN to UP  
#>  8 S1 HHGE 2000-01-01 06:25:48 DN to UP  
#>  9 S1 HGE  2000-01-01 06:25:48 DN to UP  
#> 10 S1 RGE  2000-01-01 06:25:48 UP to DN  
#> # ℹ 110,329 more rows
berth_events <- wrangle_centrix(raw_centrix, asset_map)
berth_events
#> # A tibble: 6,952 × 15
#>    signal berth train_id aspect t_enters            t_red_on           
#>    <chr>  <chr>    <int> <fct>  <dttm>              <dttm>             
#>  1 S1     A            1 G      2000-01-01 13:07:11 2000-01-01 13:07:12
#>  2 S2     B            1 G      2000-01-01 13:09:18 2000-01-01 13:09:19
#>  3 S3     C            1 G      2000-01-01 13:09:58 2000-01-01 13:09:59
#>  4 S4     D            1 G      2000-01-01 13:10:41 2000-01-01 13:10:42
#>  5 S5     E            1 G      2000-01-01 13:12:39 2000-01-01 13:12:39
#>  6 S6     F            1 G      2000-01-01 13:13:18 2000-01-01 13:13:19
#>  7 S7     G            1 G      2000-01-01 13:15:35 2000-01-01 13:15:35
#>  8 S8     H            1 G      2000-01-01 13:16:19 2000-01-01 13:16:20
#>  9 S1     A            2 G      2000-01-01 13:11:58 2000-01-01 13:11:59
#> 10 S2     B            2 YY     2000-01-01 13:12:43 2000-01-01 13:12:43
#> # ℹ 6,942 more rows
#> # ℹ 9 more variables: t_enters_next <dttm>, t_vacates <dttm>, t_red_off <dttm>,
#> #   TSAR <dbl>, T_onset <dbl>, T_clear <dbl>, T_offset <dbl>, T_travel <dbl>,
#> #   T_coach <dbl>