Wrangle Timetable Data

A wrapper for timetable processing functions filter_relevant_services(), filter_relevant_direction(), and find_calling_patterns().

Usage

wrangle_timetable(timetable, stations, stopping_stations)

Arguments

timetable

Data frame containing timetable data. Strict input validation is performed. Must be a data frame with the following columns:

train_headercharacter() Train identifier

dt_originlubridate::POSIXct() Datetime at which the train originated

geocharacter() Name of event location

eventcharacter() Type of event, must only contain the following values: Pass, Arrive, Depart, Originate, Terminate

wttlubridate::POSIXct() Scheduled datetime of event

tlubridate::POSIXct() Actual datetime of event

delaynumeric() Difference between actual time and scheduled time of event

allownumeric() Timetabled delay allowance

stations

You should provide a list of stations that you are interested in. For example, a train may stop at many stations on its journey, but your Centrix data only covers a small subset of that journey. Any trains which do not pass through any of the stations specified will be discarded. It is important to note that timetables do not only include station stops, they also include junctions and other 'Pass' events. These locations must also be included in stations, even though they are not actually stations.

Therefore, your list should completely encompasses your Centrix data. This ensures that all relevant services are included in the processed timetable. However, you may need to ensure that no irrelevant services are accidentally included.

The order in which you define this list is important if you are using wrangle_timetable() or filter_relevant_direction(), because it specifies the direction in which trains should be travelling. This allows for the filtering of services by direction. You can avoid this by using the lower level functions directly and not using filter_relevant_direction(), but these functions do not currently come with input validation, so discretion is advised.

stopping_stations

A subset of stations that specifies which stations trains can stop at. This is used for calculating calling patterns. Therefore, any train which does not stop at any of the stations specified in stopping_stations will be labeled as a 'fast' train. Trains which do stop at these stations will be labeled with 'stopping-', with the specific stations at which they stop attached as suffixes. For example, if a train stops at 'geo10', its calling pattern will be labeled as 'stopping-geo10'.

Value

A subset of the timetable which contains trains that pass through the specified stations in the given order. Time zone is set to "UTC".

See also

Examples

data(timetable, stations, stopping_stations)
timetable
#> # A tibble: 8,958 × 11
#>    train_header dt_origin           geo    event     wtt                
#>    <chr>        <dttm>              <chr>  <chr>     <dttm>             
#>  1 168H         2000-01-01 12:05:00 geo1   Originate 2000-01-01 12:20:00
#>  2 168H         2000-01-01 12:05:00 geo18  Arrive    2000-01-01 12:23:00
#>  3 168H         2000-01-01 12:05:00 geo18  Depart    2000-01-01 12:24:00
#>  4 168H         2000-01-01 12:05:00 geo2   Arrive    2000-01-01 12:28:00
#>  5 168H         2000-01-01 12:05:00 geo2   Depart    2000-01-01 12:29:00
#>  6 168H         2000-01-01 12:05:00 geo3   Arrive    2000-01-01 12:35:30
#>  7 168H         2000-01-01 12:05:00 geo3   Depart    2000-01-01 12:36:30
#>  8 168H         2000-01-01 12:05:00 geo104 Arrive    2000-01-01 12:39:00
#>  9 168H         2000-01-01 12:05:00 geo104 Depart    2000-01-01 12:39:30
#> 10 168H         2000-01-01 12:05:00 geo4   Arrive    2000-01-01 12:41:30
#> # ℹ 8,948 more rows
#> # ℹ 6 more variables: t <dttm>, delay <dbl>, allow <dbl>, allow_perf <dbl>,
#> #   allow_path <dbl>, allow_eng <dbl>
timetable_subset <- wrangle_timetable(timetable, stations, stopping_stations)
timetable_subset
#> # A tibble: 792 × 12
#>    train_header dt_origin           geo    event  wtt                
#>    <chr>        <dttm>              <chr>  <chr>  <dttm>             
#>  1 108H         2000-01-01 12:55:00 geo5   Arrive 2000-01-01 13:34:00
#>  2 108H         2000-01-01 12:55:00 geo5   Depart 2000-01-01 13:35:30
#>  3 108H         2000-01-01 12:55:00 geo6   Pass   2000-01-01 13:37:30
#>  4 108H         2000-01-01 12:55:00 geo110 Arrive 2000-01-01 13:39:00
#>  5 108H         2000-01-01 12:55:00 geo110 Depart 2000-01-01 13:39:30
#>  6 108H         2000-01-01 12:55:00 geo111 Arrive 2000-01-01 13:42:00
#>  7 108H         2000-01-01 12:55:00 geo111 Depart 2000-01-01 13:42:30
#>  8 108H         2000-01-01 12:55:00 geo112 Arrive 2000-01-01 13:44:30
#>  9 108H         2000-01-01 12:55:00 geo112 Depart 2000-01-01 13:45:00
#> 10 108H         2000-01-01 12:55:00 geo7   Arrive 2000-01-01 13:48:00
#> # ℹ 782 more rows
#> # ℹ 7 more variables: t <dttm>, delay <dbl>, allow <dbl>, allow_perf <dbl>,
#> #   allow_path <dbl>, allow_eng <dbl>, group <chr>