Estimate dwell times

This function estimates dwell times from Centrix data

Usage

estimate_dwell_times(
  berth_events_groups,
  berth_lengths,
  station_names,
  stopping_patterns,
  a_brake,
  a_tract,
  station_berth_lengths = NULL
)

Arguments

berth_events_groups

Berth-level Centrix data

berth_lengths

Data frame with 2 columns:

berthBerth ID

LBerth length in metres

station_names

Data frame with 2 columns:

berthBerth ID

stationStation name

This provides a 1-1 mapping from berth ID to station name. The station names should correspond to those in the timetable

stopping_patterns

Which stations trains of each group stop at

a_brake

Braking capacity in m/s^2

a_tract

Acceleration in m/s^2

station_berth_lengths

Data frame with 3 columns:

stationStation name

L1Distance from start of berth to station

L2Distance from station to end of berth

If station_berth_lengths is NULL (the default), then L1 and L2 will be equal to the berth length halved

Value

Data frame with 21 columns:

signalSignal ID

berthBerth ID

stationStation name

train_idTrain ID

groupGroup name

T_travelBerth travel time

v_entryVelocity at entry to berth in m/s

a_brakeBraking capacity in m/s^2

L1Distance from start of berth to station

v_exitVelocity at exit of berth in m/s

a_tractAcceleration in m/s^2

L2Distance from station to end of berth

T_brakeTime spent braking

S_brakeDistance spent braking

T_const_oneTime spent moving at a constant speed before braking for station

T_move_oneTotal time moving before reaching station

T_tractTime spent accelerating

S_tractDistance spent accelerating

T_const_twoTime spent moving at a constant speed before vacating berth

T_move_twoTotal time moving after leaving station

T_dwellDwell time at station