Preprocess raw Centrix data

preprocess_signal_events() handles only signals, defined by raw_centrix$asset starting with the character 'S'.

preprocess_track_events() handles only tracks, defined by raw_centrix$asset starting with the character 'T'.

Usage

preprocess_signal_events(raw_centrix, asset_map, state_map)

preprocess_track_events(raw_centrix, asset_map)

Arguments

raw_centrix

A data frame containing raw Centrix data. Strict input validation is applied to ensure the data adhere to the expected structure. The data should be in a data frame with the following columns:

asset: A character() vector containing the asset ID. Tracks are expected to follow the regex pattern "T[A-Z]+(-[0-9])?". E.g., 'TABC' 'TABC-1' 'TABC-2'. Signals are expected to follow the regex pattern "S[0-9]+\s[A-Z]+". E.g., 'S123 HGE' 'S123 I'. The second part of the signal ID should be a state code referring to the signal aspect, e.g. 'RGE' for red or 'DGE' for green. For other codes see: https://wiki.openraildata.com/index.php/Signalling_Nomenclature#Signals.

dt: A lubridate::POSIXct() object representing the date and time at which the observation was made.

transition: A character() vector representing the state transition that caused the observation. A transition can be one of "UP to DN" or "DN to UP", anything else will be ignored. For tracks, "UP to DN" signifies a train entering the track, while "DN to UP" is a train vacating the track. For signals, "UP to DN" represents an aspect change. Therefore, any signals with a "DN to UP" transition will be ignored.

asset_map

The asset map is a representation of the track section. Currently, the track is represented by a linear data frame. However, this does not support more complex track layouts, such as junctions. The asset map should be a data frame containing a 1-1 mapping from signal ID to berth name, a 1-many mapping from berth name to track ID, and an extra event column. The event column represents the fact that there are separate Centrix observations for trains entering and exiting a track. Therefore, each track can have both an 'enters' and a 'vacates' events associated with it. Note that in order to calculate berth travel times, it is necessary to know when the train entered to next berth. Therefore, it may be impossible to calculate travel times for the final berth. The asset map data frame should adhere to the following structure:

signal: A character() vector containing the signal ID. It should follow the regex pattern "S[0-9]+".

berth: A character() vector containing the berth ID. It should follow the regex pattern "[A-Z]+".

track: A character() vector containing the track ID. It should follow the regex pattern "T[A-Z]+(-[0-9])?".

event: A character() vector containing the event type. It must be one of either "enters" or "vacates".

state_map

The state mapping provides a 1-1 mapping from signal state codes to signal aspects. This is standardised, so railvarr provides a default mapping (see state_mapping). Simply leave this parameter blank to make use of this default mapping. See this website for other codes.

Value

preprocess_signal_events returns a data frame containing aspect events with columns:

signal character() signal ID.

dt lubridate::POSIXct() datetime.

aspect factor() signal aspect caused by the event.

past_aspect factor() signal aspect preceding the event.

preprocess_track_events returns a data frame containing track events with columns:

track character() track ID.

dt lubridate::POSIXct() datetime.

occcupied logical() TRUE if train enters track, else FALSE.

event character() 'enters' if train enters track, else 'vacates'.

Details

preprocess_signal_events() converts raw_centrix$asset into signal ID and signal state, the latter of which is converted to signal aspect using the state_map..

preprocess_track_events() converts raw_centrix$transition into track entry and exit events.

See also

Examples

data(raw_centrix, asset_map, state_mapping)
raw_centrix
#> # A tibble: 110,339 × 3
#>    asset   dt                  transition
#>    <chr>   <dttm>              <chr>     
#>  1 TA-1 TR 2000-01-01 06:13:48 UP to DN  
#>  2 S1 HHGE 2000-01-01 06:13:49 UP to DN  
#>  3 S1 HGE  2000-01-01 06:13:49 UP to DN  
#>  4 S1 RGE  2000-01-01 06:13:49 DN to UP  
#>  5 TA-1 TR 2000-01-01 06:25:29 DN to UP  
#>  6 TA-1 TR 2000-01-01 06:25:31 UP to DN  
#>  7 TA-1 TR 2000-01-01 06:25:43 DN to UP  
#>  8 S1 HHGE 2000-01-01 06:25:48 DN to UP  
#>  9 S1 HGE  2000-01-01 06:25:48 DN to UP  
#> 10 S1 RGE  2000-01-01 06:25:48 UP to DN  
#> # ℹ 110,329 more rows
aspect_events <- preprocess_signal_events(raw_centrix,
                                          asset_map,
                                          state_mapping)
aspect_events
#> # A tibble: 32,512 × 4
#>    signal dt                  aspect past_aspect
#>    <chr>  <dttm>              <fct>  <fct>      
#>  1 S1     2000-01-01 06:13:49 R      NA         
#>  2 S1     2000-01-01 06:25:48 YY     R          
#>  3 S1     2000-01-01 06:25:48 Y      YY         
#>  4 S1     2000-01-01 06:25:53 R      Y          
#>  5 S1     2000-01-01 06:26:16 YY     R          
#>  6 S1     2000-01-01 06:26:16 Y      YY         
#>  7 S1     2000-01-01 07:44:16 R      Y          
#>  8 S1     2000-01-01 08:06:31 YY     R          
#>  9 S1     2000-01-01 08:06:31 Y      YY         
#> 10 S1     2000-01-01 08:20:08 R      Y          
#> # ℹ 32,502 more rows

#----------------------------------------------------------------------------

data(raw_centrix, asset_map)
raw_centrix
#> # A tibble: 110,339 × 3
#>    asset   dt                  transition
#>    <chr>   <dttm>              <chr>     
#>  1 TA-1 TR 2000-01-01 06:13:48 UP to DN  
#>  2 S1 HHGE 2000-01-01 06:13:49 UP to DN  
#>  3 S1 HGE  2000-01-01 06:13:49 UP to DN  
#>  4 S1 RGE  2000-01-01 06:13:49 DN to UP  
#>  5 TA-1 TR 2000-01-01 06:25:29 DN to UP  
#>  6 TA-1 TR 2000-01-01 06:25:31 UP to DN  
#>  7 TA-1 TR 2000-01-01 06:25:43 DN to UP  
#>  8 S1 HHGE 2000-01-01 06:25:48 DN to UP  
#>  9 S1 HGE  2000-01-01 06:25:48 DN to UP  
#> 10 S1 RGE  2000-01-01 06:25:48 UP to DN  
#> # ℹ 110,329 more rows
track_events <- preprocess_track_events(raw_centrix,
                                        asset_map)
track_events
#> # A tibble: 25,242 × 4
#>    track dt                  occupied event  
#>    <chr> <dttm>              <lgl>    <chr>  
#>  1 TA-1  2000-01-01 06:13:48 TRUE     enters 
#>  2 TA-1  2000-01-01 06:25:29 FALSE    vacates
#>  3 TA-1  2000-01-01 06:25:31 TRUE     enters 
#>  4 TA-1  2000-01-01 06:25:43 FALSE    vacates
#>  5 TA-1  2000-01-01 06:25:52 TRUE     enters 
#>  6 TA-1  2000-01-01 06:25:54 FALSE    vacates
#>  7 TA-1  2000-01-01 06:25:57 TRUE     enters 
#>  8 TA-1  2000-01-01 06:25:59 FALSE    vacates
#>  9 TA-1  2000-01-01 06:26:02 TRUE     enters 
#> 10 TA-1  2000-01-01 06:26:11 FALSE    vacates
#> # ℹ 25,232 more rows